requirements.txt
opencv-python>=4.10
numpy>=1.23
scikit-image>=0.22
scikit-learn>=1.4
matplotlib>=3.8
seaborn>=0.13
torch>=2.2
torchvision>=0.17
fastapi>=0.104.0
uvicorn>=0.23.0
python-multipart>=0.0.6
pillow>=10.0.0
python-dotenv>=1.0.0
